chore(test): run kubernetes tests on linux platform in github actions Signed-off-by: Anton Misskii <amisskii@redhat.com>

+++ tests/playwright/src/specs/deploy-to-kubernetes.spec.ts
@@ -19,6 +19,7 @@
 import { ContainerState } from '../model/core/states';
 import type { ContainerInteractiveParams } from '../model/core/types';
 import { ContainerDetailsPage } from '../model/pages/container-details-page';
+import { KubeContextPage } from '../model/pages/kubernetes-context-page';
 import { expect as playExpect, test } from '../utility/fixtures';
 import {
   createKindCluster,
@@ -42,7 +43,7 @@ const CONTAINER_START_PARAMS: ContainerInteractiveParams = {
   attachTerminal: false,
 };
 
-const skipKindInstallation = process.env.SKIP_KIND_INSTALL ? process.env.SKIP_KIND_INSTALL : false;
+const skipKindInstallation = process.env.SKIP_KIND_INSTALL === 'true';
 
 test.beforeAll(async ({ runner, welcomePage, page, navigationBar }) => {
   test.setTimeout(350_000);
@@ -55,26 +56,30 @@ test.beforeAll(async ({ runner, welcomePage, page, navigationBar }) => {
     await settingsBar.cliToolsTab.click();
 
     await ensureCliInstalled(page, 'Kind');
+    await createKindCluster(page, CLUSTER_NAME, true, CLUSTER_CREATION_TIMEOUT);
   }
-  await createKindCluster(page, CLUSTER_NAME, true, CLUSTER_CREATION_TIMEOUT);
+  if (process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux') {
+    const settingsBar = await navigationBar.openSettings();
+    const kubePage = await settingsBar.openTabPage(KubeContextPage);
+    await playExpect(kubePage.heading).toBeVisible();
+
+    playExpect(await kubePage.isContextDefault('kind-kind-cluster')).toBeTruthy();
+  }
 });
 
 test.afterAll(async ({ runner, page }) => {
   test.setTimeout(90000);
   try {
-    await deleteContainer(page, CONTAINER_NAME);
-    await deleteImage(page, IMAGE_TO_PULL);
-    await deleteKindCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
+    if (!(process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux')) {
+      await deleteContainer(page, CONTAINER_NAME);
+      await deleteImage(page, IMAGE_TO_PULL);
+      await deleteKindCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
+    }
   } finally {
     await runner.close();
     console.log('Runner closed');
   }
 });
-
-test.skip(
-  !!process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux',
-  'Tests suite should not run on Linux platform',
-);
 
 test.describe.serial('Deploy a container to the Kind cluster', { tag: '@k8s_e2e' }, () => {
   test('Pull an image and start a container', async ({ navigationBar }) => {
